:recycle: refactor(api): move ItemXxx -> ItemLocationXxx
