+++ backend/streetdrop-api/src/main/java/com/depromeet/domains/item/repository/ItemLocationRepository.java
@@ -1,62 +1,18 @@
 package com.depromeet.domains.item.repository;
 
-import com.depromeet.domains.item.dao.ItemPointDao;
-import com.depromeet.domains.user.dao.UserItemPointDao;
-import com.querydsl.core.types.Projections;
-import com.querydsl.jpa.impl.JPAQueryFactory;
-import lombok.RequiredArgsConstructor;
-import org.locationtech.jts.geom.Point;
-import org.springframework.stereotype.Repository;
+import com.depromeet.item.ItemLocation;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 
-import java.util.List;
+public interface ItemLocationRepository extends JpaRepository<ItemLocation, Long> {
 
-import static com.depromeet.external.querydsl.mysql.spatial.MySqlSpatialFunction.mySqlDistanceSphereFunction;
-import static com.depromeet.item.QItem.item;
-import static com.depromeet.item.QItemLocation.itemLocation;
-import static com.depromeet.music.album.QAlbumCover.albumCover;
+    @Query("SELECT count(il) FROM ItemLocation il JOIN FETCH il.villageArea va WHERE va.cityArea.cityName = :city")
+    Integer countItemsByCity(String city);
 
-@Repository
-@RequiredArgsConstructor
-public class ItemLocationRepository {
+    @Query("SELECT count(il) FROM ItemLocation il JOIN FETCH il.villageArea va WHERE va.cityArea.stateArea.stateName = :state")
+    Integer countItemsByState(String state);
 
-    private final JPAQueryFactory queryFactory;
-
-    public List<ItemPointDao> findNearItemsPointsByDistance(Point point, Double distance, Double innerDistance, List<Long> blockedUserIds) {
-        return queryFactory.select(
-                        Projections.fields(
-                                ItemPointDao.class,
-                                itemLocation.point,
-                                item.id,
-                                albumCover.albumThumbnail,
-                                mySqlDistanceSphereFunction(itemLocation.point, point)
-                                        .loe(String.valueOf(innerDistance))
-                                        .as("isInnerDistance")
-                        ))
-                .from(itemLocation)
-                .join(itemLocation.item, item)
-                .on(itemLocation.item.id.eq(item.id))
-                .join(itemLocation.item.albumCover, albumCover)
-                .on(item.albumCover.id.eq(albumCover.id))
-                .where(mySqlDistanceSphereFunction(itemLocation.point, point).loe(String.valueOf(distance)))
-                .where(itemLocation.item.user.id.notIn(blockedUserIds))
-                .fetch();
-    }
-
-    public List<UserItemPointDao> findUserDropItemsPoints(Long userId) {
-        return queryFactory.select(
-                        Projections.fields(
-                                UserItemPointDao.class,
-                                itemLocation.point,
-                                item.id,
-                                albumCover.albumThumbnail
-                        ))
-                .from(itemLocation)
-                .join(itemLocation.item, item)
-                .on(itemLocation.item.id.eq(item.id))
-                .join(itemLocation.item.albumCover, albumCover)
-                .on(item.albumCover.id.eq(albumCover.id))
-                .where(item.user.id.eq(userId))
-                .fetch();
-    }
+    @Query("SELECT count(il) FROM ItemLocation il")
+    Integer countAll();
 
 }
