+++ backend/streetdrop-api/src/main/java/com/depromeet/domains/item/repository/ItemRepository.java
@@ -24,12 +24,4 @@ public interface ItemRepository extends JpaRepository<Item, Long>, QueryDslItemR
 
     Long countByUser(User user);
 
-    @Query("SELECT count(i) FROM Item i JOIN FETCH i.itemLocation il JOIN FETCH il.villageArea va WHERE va.cityArea.cityName = :city")
-    Integer countItemsByCity(String city);
-
-    @Query("SELECT count(i) FROM Item i JOIN FETCH i.itemLocation il JOIN FETCH il.villageArea va WHERE va.cityArea.stateArea.stateName = :state")
-    Integer countItemsByState(String state);
-
-    @Query("SELECT count(i) FROM Item i")
-    Integer countAll();
 }
